Working on Windows

+++ interval_file_roll_trigger.cpp
@@ -33,27 +33,27 @@ void interval_file_roll_trigger::compute_next_roll()
     if (period_ == period::MINUTE)
     {
         pcs.tm_sec = 0;
-        pcs.tm_min += count_;
+        pcs.tm_min += static_cast<int>(count_);
     }
     else if (period_ == period::HOUR)
     {
         pcs.tm_sec = 0;
         pcs.tm_min = 0;
-        pcs.tm_hour += count_;
+        pcs.tm_hour += static_cast<int>(count_);
     }
     else if (period_ == period::DAY)
     {
         pcs.tm_sec = 0;
         pcs.tm_min = 0;
         pcs.tm_hour = 0;
-        pcs.tm_mday += count_;
+        pcs.tm_mday += static_cast<int>(count_);
     }
     else if (period_ == period::WEEK)
     {
         pcs.tm_sec = 0;
         pcs.tm_min = 0;
         pcs.tm_hour = 0;
-        pcs.tm_mday += count_ * 7;
+        pcs.tm_mday += static_cast<int>(count_ * 7);
     }
     else if (period_ == period::MONTH)
     {
@@ -61,7 +61,7 @@ void interval_file_roll_trigger::compute_next_roll()
         pcs.tm_min = 0;
         pcs.tm_hour = 0;
         pcs.tm_mday = 1;
-        pcs.tm_mon += count_;
+        pcs.tm_mon += static_cast<int>(count_);
     }
     next_roll_ = clock_type::from_time_t(calendar::to_time_t(pcs));
 }
